+++ ttnn/cpp/ttnn/operations/conv/conv2d/prepare_conv2d_weights.hpp
@@ -114,26 +114,6 @@ ttnn::Tensor prepare_conv_bias(
     const std::optional<const DeviceComputeKernelConfig>& compute_config_,
     const std::optional<const Conv2dSliceConfig>& dram_slice_config_ = std::nullopt);
 
-Conv2dSliceConfig determine_dram_slice_config_for_prepare_weights(
-    uint32_t batch_size,
-    uint32_t input_height,
-    uint32_t input_width,
-    uint32_t in_channels,
-    uint32_t out_channels,
-    std::array<uint32_t, 2> kernel_size,
-    std::array<uint32_t, 2> stride,
-    std::array<uint32_t, 4> padding_n4,
-    std::array<uint32_t, 2> dilation,
-    uint32_t groups,
-    Layout input_layout,
-    DataType input_dtype,
-    DataType conv_output_dtype,
-    bool has_bias,
-    const Conv2dConfig& conv_config,
-    const DeviceComputeKernelConfig& compute_config,
-    MeshDevice* device,
-    std::optional<Conv2dSliceConfig> input_dram_slice_config = std::nullopt);
-
 // Unified parameter struct for conv2d weight and bias preparation
 struct Conv2dWeightsBiasPrepConfig {
     // Constructor to ensure all required parameters are initialized
